Fix get_server_name method (#2)
Task/Issue URL: https://app.asana.com/0/488551667048375/1203126514402185/f

### Description
When parsing the ClientHello header we made a mistake when advancing through the extensions. We should increase the offset by `extension_len` however we increased the offset by `extensions_len`–which is the length of all extensions.

As we were looking for one particular extention, ie. "server name"–type=0x0000– the crash would happen when the "server name" extension was not the first one.

Incidentally added:
* PR template
* `.gitignore` file
* `CMakeList` that currently just lists the files but doesn't compile. This is just to little by little prep the library to be developed in isolation

### Steps to test this PR
- [x] In the android repo `cd ~/code/ddg/Android/vpn-network/vpn-network-impl/src/main/cpp/netguard`
- [x] fetch this PR branch, ie. `fix/aitor/atp/get_sni_server_name`
- [x] go back to the Android project root directory
- [x] run `git submodule` and ensure the `vpn-network/vpn-network-impl/src/main/cpp/netguard` submodule points to `ebc033f7c54b3dd32299c189e8b1c0144d39819a` commit, ie. `heads/fix/aitor/atp/get_sni_server_name` branch
- [x] build the android app
- [x] launch app and enable AppTP
- [ ] ensure AppTP blocks as usual when launching apps that attempt to track, ie. speedtest